requirements.txt
yfinance
pandas
apache-airflow==2.9.3
psycopg2-binary